Fix error return value from RecordsArray
It should not return a double nested array for errors.
This can be interpreted as one empty row and breaks code like
  if (!$result)
because it evaluates to true because the array has one element.
For errors it should simple return an empty array, i.e. no rows.
(I would prefer false but I have no reason to change it more)